Add the "examples" keyword.
This addresses issue #121.

Most of the wording is lifted from "default", which is similar and can serve
the same purpose.  This wording explicitly allows for using "default" as an
example, without requiring that implementations do so.

The "MAY be used in a root schema or any subschema" line for all of the
meta-data keywords seems to have been a holdover from an earlier approach
of always specifying where keywords could be used.  It is not done anywhere
else and raised a red flag in review of the original version of this
commit, so this version removes it from all meta-data keywords.