Futher clarify primitive vs child validation.
A lot of the most confusing language around the "additional" keywords
can be removed if we consider keywords strictly with respect to either
the instance's primitive type or against child instances.

Describing {"additionalItems": false} as failing child validation of
elements beyond the length of "items" is more straightforward and
has the same effect as the more convoluted reasoning about length.