"uritemplate" and "jsonpointer" formats
Addresses issue #109 by adding "uritemplate" and "jsonpointer",
as defined by their RFCs, both of which already play roles
within JSON (Hyper-)Schema.

We use "uritemplate" in hyper-schema, but until the preprocessing
step is removed we technically can't put the format in the
meta-hyperschema.  This is being addressed elsewhere.

While we do not need "jsonpointer" for our meta-schema, it is
used for schema+json URI fragments, and in other JSON-based
media types such as JSON Patch.