Define subschemas -> child instances in core
This is a trial balloon for fixing #161.

This moves several keywords that control which subschemas apply
to which child instances from validation to core.  The only
actual validation that these keywords defined was that all child
instances need to validate against all applicable subschemas.

This simple child validation principle has been moved up to
the general considerations section of the validation spec.

The moved keywords are now prefixed with "$", which is what we've
discussed doing for all core keywords (Another potential core
keyword would be "$data", for instance, as it would be used in
multiple vocabularies depending on which proposals we accept,
if any.)

As with "id" vs "$id" and the boolean forms of
exclusiveMinimum/Maximum, the old forms are still allowed but
expected to be removed before RFC.  Migrating an old schema
to the new form is trivial, and was discussed before we decided
to rename "id".  It is equally applicable here.