Remove confusing use of "absolute" URI.
The value of "$schema" is not required to be an absolute URI
as defined by RFC 3986 (with a scheme, but no fragment).  It just
needs to be a URI and not a relative reference.

This is particularly important as the meta-schemas include
an empty fragment in their id URIs, which would be illegal if
RFC 3986 section 4.3 Absolute URIs were required.