Restore exclusive* dependencies, update metaschema
This restores dependencies that got dropped somewhere between
draft 04 and 05 (not clear where).

It adds a reasonable restriction that the boolean vs numeric
forms not be mixed for exclusiveMinimum and exclusiveMaximum,
and implements the correct meta-scheam for this.  Which needs
a oneOf to corectly group the dependencies, but no longer needs
an extra allOf or anyOf to allow minimum and maximum to change
types independently.