Wording clarifications in general considerations.
These are the leftover bits of Issue #55 and some clarifications
requested in a comment on issue #101 that have not already been
added in some other PR for some other issue.

These specific chagnes were previously approved in #143, but so
many other things have changed since #143 that most of it was
no longer relevant, so I closed it and started these changes over.

In particular, explaining {} and {"not": {}} is no longer needed
as they are covered while introducing "true" and "false" schemas
in the core specification, so that is no longer repeated in this
change.

Likewise, the parent/child validation descriptions have been
modified in several PRs and no longer has the problems that were
previously a concern.