Standardize on hyphenated format names
When we had ore formats in earlier drafts, all compound names
(such as "date-time") were hyphenated.  Draft 05 added "uriref"
without a hyphen, which started looking weird when we added
"uritemplate" and "jsonpointer" without hyphens to match it.
This makes everything hyphenated (note "hostname" was deemed
to be a single word in this context).

"uriref" became "uri-reference" as that is the form that is
most common in other RFCs.