Another attempt at clarifying objects/arrays.
This goes back to Draft 04's approach of grouping keywords,
but mostly retaining Draft 05's more concise wording and
per-keyword organization.

The general considerations section on validating primitive types
and child values has been simplified, and the details have been
moved down to each grouping of array or object keywords.  This
puts the information near where it is most relevant, which
allows removing some of the more confusing working in the
keywords that apply to child values.