Clearly define schema+json fragid forms.
Reference the JSON Pointer RFC for using JSON Pointers
as fragids (includes encoding requirements).

Remove the syntax description for how to write a fragment
URI reference (leading "#") as that is covered by RFC 3986.

Note that it is possible to give a fragment that is not
valid in either style of fragid, and document that the
effect of such a fragment is undefined.

Also fill in a bit more of the IANA considerations section.