Provide guidance on hyperschema and validation
It's been commonly assumed that hyper-schema keywords apply only
when validation is successful.  Be explicit about this, and resolve
ambiguous situations.

Most notably, remove any burden on implementations to figure out
what to do with situations such as:

{"not": {"not": {"links": [...]}}}

where the schema containing links technically must validate against
the instance, but any less trivial nesting of "not"s
(including such situations as a "not" inside of a "oneOf" branch
that fails validation while the overall "oneOf" succeeds) rapidly
becomes extremely difficult to reason about.