Remove leftover wording from preprocessing.
We removed the preprocessing rules, including those related to
using the entire instance value, and made a note of the limitation
and intent to remove it in a future draft.

This stray phrase should have been removed at that time.

Also add a cref acknowledging the limitations and further work
needed in this area.  There was such a cref in the previous
draft and I had thought I'd updated it rather than deleting it,
but apparently not.  It was a good idea, let's keep it.