"method" does not directly constrain HTTP methods
"method" already documented that "method": "get" does not constrain
the LDO to GET.  Clarify that "post" also does not directly constrain
the LDO to HTTP POST, although the semantics of "schema" and "encType"
may have that effect in practice as they MUST be used for the
request payload if present.

Aside from that indirect constraint, an LDO can be used with multiple methods.

To clarify non-HTTP usage in generic terms, also add  some wording
about request formats being defined in terms of the target
representation or not.  This should provide more guidance on how
to use both "schema" and "targetSchema", independent of URI scheme.