"method" does not constrain HTTP methods
"method" already documented that "method": "get" does not constrain
the LDO to GET.  Clarify that "post" also does not constrain the
LDO to HTTP POST, and that the LDO can be used with multiple methods.
Further clarify atht "schema" SHOULD be ignored if it is not relevant.

The "schema" and "targetSchema" keywords are used with relevant
HTTP methods regardless of the presence, absence, or value of
JSON Hyper-Schema LDO "method".  This is consistent with the
existing wording of both "schema" and "targetSchema" around the
construction of PUT requests.

To clarify non-HTTP usage in generic terms, add some wording
about request formats being defined in terms of the
target representation or not.  This should provide more
guidance for when "method": "post" and "schema" are relevant,
independent of URI scheme.