Remove "method", add "form".
This removes the "method" keyword and reworks all references to
HTML forms.  The difference in approach between JSON Hyper-Schema
and HTML+JavaScript is explained, and the different sorts of links
(in terms of data usage) are described.

The "form" keyword is being added to indicate HTML POST form
semantics.  GET forms can be simulated with "hrefSchema", and the
guidance on that has been made more clear.

We also now clearly acknowldege that you cannot directly specify
HTML methods, as that was the largest source of confusion in the
feedback to the previous draft.