Split validation keywords into sections.
This improves the readability of the spec, particularly
new readers trying to get a feel for the scope from the
table of contents.

A previous, more complex reorganization that was attempted
before got plenty of support, and this more minimal change
avoids the issues that bogged the previous approach down.

Finally, having sections will make the section grouping
"if", "then", and "else" make more sense.  Adding those
keywords without grouping the would seem odd.