Add "$comment" as a keyword.
This addresses issue #197.  The primary purpose is to reserve the
keyword and ensure that implementations do NOT do anything based
on its presence or contents.  This improves interoperability by
ensuring that comments are safe for use with any conforming system.

Without this keyword, two implementations may implement it with
differing semantics, potentially producing undesirable or
insecure behavior.

Effort has been made to ensure that existing implementations
that properly ignore unknown keywords will not need to make
any changes at all to be in conformance for this keyword.