Start a contributing guide from the wiki contents.
This file has exactly the contents of the front wiki page, except
for the I-D and meta-schema publishing processes which are not
relevant to most contributors.

I'm moving it into a CONTRIBUTING.md as it integrates with
GitHub's tools (e.g. auto-linking it to the create PR screen)
and will allow us to discuss changes as PRs.

I will change the front wiki page to link to this (but retain
the sections that have not been moved here).