Reference validation from hyper-schema
Despite having a whole section describing how hyper-schema
builds on validation, it was never referenced or listed
as a normative specifciation.

Also, we had labled the meta-schema link as a validation schema
in an apparent copy-paste error.

Finally, improve wording around building on the other specs in
the same way as was done for the validation spec referincing core.