What is a JSON Hyper-Schema?
This reworks the general descriptive text to emphasize how
Hyper-Schema fits in to the RFC 5988 link model, or more
accurately the RFC 5988bis revision currently in development.

By presenting the Web Linking model as the central foundation
for Hyper-Schema, and emphasizing the relationship between
Hyper-Schema and the link context, the focus should be a bit
more clearly on hypermedia rather than api request/responses.

Also defined "operation" which I apparently just started using
in that section without explaining it at all.