Move "readOnly" from Hyper-Schema to Validation
We move the description and schema definition of "readOnly" verbatim
from the Hyper-Schema specification (and meta-schema) to the validation
one, in the "Metadata keywords" section (along with "default", in
particular).

This addresses part of #363, which states that some keywords
historically in the Hyper-Schema specification would be better in the
Validation document. The main argument for moving keywords such as
"readOnly" is that the Hyper-Schema document could then only focus on
describing the hypermedia linking model.