Add "targetHints"
This is a very tentative first pass to get the discussion started.

The goal for this draft is to specify this well enough to get
real-world implementations and feedback.  It seems best to
approach that goal by providing loose guidelines and encouraging
experimentation.

The lack of a standard way to advertise HTTP method support,
for instance, is a major barrier to adoption for many who have
expressed interest in Hyper-Schema.  Including an approach that
will easily fulfill that need while encouraging feedback on
what further guidance is needed for more comprehensive use
is preferable to waiting another draft while we come up with
a solid proposal.  That still might not hold up in the real
world anyway.

This will be more useful for feedback than simply providing
an "allow" keyword, as it will encourage exploring other
use cases, which a narrowly targeted "allow" keyword would not.