Add "deprecated" and "patternDeprecated"
Addresses #74, specifically deprecating object members.

Originally proposed as a field within the schema of the
thing to be deprecated, that arrangement produced numerous
situations with unclear semantics.

Later, @awwright proposed that we look to "required" for
the proper model, declaring the deprecation at the object
level separate from the property schemas.

This proposal allows some flexiblity in how much information
is conveyed about the deprecation, which seems useful given
the broad range of possible applications.

This is being added to the meta-data/annotation section of
the validation spec for the same reason that "readOnly"
and "media" are being moved over.  Yes, it is unclear whether
meta-data/annotation keywords belong in validation at all,
but please do not use this PR for that discussion.

Deprecating entire resources is to be considered separately,
and is probably best handled for HTTP by a "targetHint" of the
proposed "Sunset" header.

It is not immediately clear to me whether there is a need for
"additionalDeprecated", so I left it out.