Restore "regex" to the list of formats.
Addresses #392.  It's come up several times that removing
this format did not make sense and was quite possibly an accident.

The "regex" format was actually still in use in the meta-schema
for "pattern".  I've added it to "patternProperties" as well
using "propertyNames".

Regular expressions are a fundamental programming tool, and
play a part in JSON Schema's specification.  It seems incorrect
to not offer semantic documentation / optional validation of them.