Update ipv6 RFC reference.
RFC 2373 has long since been obsoleted, and its replacement has
likewise been obsoleted.  RFC 4291 is the current RFC.  While there
is a further RFC providing recommendations for consistent formatting
of ipv6 addresses, it includes the requirement that all implementations
MUST support the full syntax laid out in RFC 4291, so that is what
we give as the reference.