Note that we may remove "dependencies"
This was discussed in the if/then/else issue.  There was significant
support for removing "dependencies" as it is more straightforward,
although sometimes more verbose, to implement the same funcitonality
with if/then/else.

Rather than going straight to ripping it out, let's give if/then/else
a bit of time to gather feedback.  Make a note of it to warn that
"dependencies" may be removed, which will hopefully provoke some
feedback on that idea as well.