Remove duplication and descriptions.
We regularly forget about links.json.  Let's use $ref and prevent
that.

Also, remove all of the descriptions.  They are not present in the
core/validation schema and have gotten out of sync several times.
The descriptions tend to be either trivial or inadequate as the
more complex features aren't easily described in a readable amount
of JSON text.

This addresses #175