Minimal change from "profile" to "schema"
This change removes "profile", since as noted in issue #222 the
usage is not correct.  It replaces it with "schema", with the
minimal possible change.

We may wish to also remove the "describedby" link relation and
instead mandate using "schema", or a similarly named
extension relation, and allow downloading schemas based on the
parameter.  That is outside of the scope of this change.