Allow "headerSchema" for content negotiation
While reworking the spec for overall flow, it became apparent
that a valuable use of "headerSchema" is to advertise content
negotiation options.  However, as currently worded, this usage
is forbidden.

Change the usage to allow this without being too burdensome
on implementations.

The rewrite will provide greater clarity on the general
implementation requirements of "targetHints" and "headerSchema",
so please do not worry too much about that for now.