Add $vocabularies, clarify $schema and $ref
$schema is now explicitly intended for meta-schema declaration.
An explanation for why it MUST only be present in root schemas
has been added so we (I) don't forget about it and freak out (again).

The concept of vocabularies is now introduced explicitly, and
the $vocabularies keyword is introduced to declare vocabulary
support.  For compatibility (and simplicity in the case of
a single standard meta-schema conveying sufficient information),
omitting $vocabularies in the root schema causes it to behave
as if the $schema value is listed as a vocabulary.  This preserves
all existing behavior.

Finally, a paragraph is added to $ref clarifying the conceptual
model, specifically with respect to $schema and meta-schema
validation.