Reconstruct rel-ptr draft XML
This is the best I can manage for reproducing the previous
draft's XML (but adding myself as editor and updating the draft name).
A few of the xref's may be in slightly different styles, as I've
never entirely figured out all of the options for that.  But they
are all equally readable compared to the past generated document
as far as I can tell.