Trim down JSON Hyper-Schemas in APIs appendix
A lot of this should go on some sort of best practices guide
on the web site.  This cuts quite a bit of text, and merges
the static analysis appendix into the API appendex as API-wide
processing is the main use case for such analysis.

There is likely more to cut, and perhaps the entire appendix
should be removed, but I'd like to get some feedback on this
level of information first.