Add output format schema, tidy up other schemas
Also improve the wording introducing the output format.

Split the "links" fields off from their "required" value, as
we may use the format in environments that do not require the same
set of fields.  For instance, the output format is allowed to omit
"href", which is required in an LDO proper.

Finally, update the URIs to draft-07-wip which is less confusing
than having them say draft-06 (which is blatantly wrong) or
draft-07 (which is misleading as it's not yet finalized, much
less published).