New overview based on applicability and assertions
This reworks the "General validation considerations" section into
an overview that incorporates the applicability concept that was
recently added to the annotation section.  Some of the content
from that section was removed as it became duplicate content
with the new Overview.

The new Overview also talks about assertions, which is a word
used in the existing Abstract and Introduction sections.  It
helps by providing a name for the non-annotation keyword behavior,
other than "validation" which is used for the complete
specification.

The very awkward "Validation of primitive types and child values"
is covered much more clearly when written in terms of applicability
vs assertions.  These concepts map naturally to the distinction
that was being made in that section, and make it less of a weird
exception that needs explaining.

The other old sections are also covered in the new wording,
but are less dramatically affected.  And in some cases, kept
nearly word-for-word in new locations.