Clarify data model vs unicode characters
JSON Schema does not care about unicode characters, only the
data model.  Vocabularies SHOULD account for this by working
with data model strings instead of the original unicode
representation of numbers.

For example, define a value for the "format" validation keyword
that indicates that a string should be interpreted as a fixed-point
decimal if you wish to retain precision information in your data.