Incorporate awwright's suggestions on cref #1
Better explanation of the concerns around and possible future
direction of $schema usage.

Also add an XML comment to remind us of the most notable concern,
so we can skip the whole thing where we forget about it and then
argue about it until we work it out again.  Which of course has
never happened...