Add a bit more about target attribute authority
While "targetSchema" is technically never required to process
a response, as the response should indicate its own schema,
other target attributes in the LDO, such as the submission schema,
cannot be conveyed in any way other than through the LDO.

It would seem like we need a provision for determining when the
LDO can be constructed, but it's not entirely clear how that
should work.  Put in a bit about it and a CREF to hopefully
attract the attention of someone who can improve the section.