Use boolean true instead of {} where possible
Let's emphasize the clarity of the boolean schemas as that is
why they were added.

Note that the default values of "definitions", "properties",
and "patternProperties" is an actual empty object, not an
empty schema, so those stay as they were.

We were using an explicit empty schema to emphasize/document
that a keyword accepts any possible value in several places,
but were not using it in a few others such as "items" within
the "enum" schema.  Consistently make it explicit.

There were no empty object subschemas in the other meta-schemas.