Fix several early changelog oddities
The core spec's change log seemed to have been spliced across
draft-04 and draft-00.  This dates back to draft-04.  Move the
obviously draft-04-related stuff up to that part of the changelog,
and add a line about splitting out hypermedia (analogous to the
existing line about splitting out validation).

It did not seem worthwhile to import the old changelogs from
draft-01, -02, and -03, although we can do that if necessary.
This change just fixes things that really did not make sense.

Also, in both the validation and hyper-schema specs, the oldest
draft was erroneously numbered -01 when it should be -00.