Add : to hanging list lables w short bodies
The main HTML rendering of the spec on the IETF site (and the
plain text rendering) put the first line of a hanging text
body on the same line as the label.  When the bodies are each
a single line, this looks like a weird spacing error rather
than a labeled list.  Use colons to make these lists more listy.

This is *not* done with the body is itself a list, as the plain
text formatting handles that in a way that is already clear,
and the colon looks odd particularly for the list where the
labels are JSON Pointers.