Remove "Internal Referencing" and rework example
We know from feedback that the "internal referencing" concept
and in particular the "same document" wording is confusing and
for some people misleading.  Remove the term and focus the wording
on normal RFC 3986 URI reference resolution.

Cover the same logical ground as "internal referencing" by discussing
whether and how the implementation knows of the identified schema,
using the same general ideas as covered in the previous section
on loading (or not loading) referenced schemas.