More thorough rewrite of "if"/"then"/"else"
Further feedback indicates the language was still confusing.
Rewrote it to be more explicit about how the keywords interact
with each other and with both validation and annotation
collection.

In particular, an "if" on its own collects annotations but
never affects validation, annotations are only collected from
"them" or "else" if their use is indicated by the validation
outcome of an adjacent "if", and there is no default behavior
for any of the keywords.

Technically, "then" could be assigned an empty schema default
without causing problems, but that felt weirdly unbalanced.

If "if" has a default of true/empty schema, then "then" on
its own would be active, which is counter-intuitive.  If "else"
had a default it would need to be false rather than true, and
saying it has no default seemed simpler, and matches how "not"
is handled.  "not" is the only other keyword that would otherwise
need a false schema as a default value.