Move applicability keywords and text to core
This just moves text, with the bare minimum title, link target,
and indentation changes necessary to make it fit in the core spec.

The Applicability overview section becomes the intro to the
applicability vocabulary, which is followed by the keywords.

Further commits will rework the wording for the new arrangement
and probably further re-organize the applicability overview.

Note that "dependencies" is left in the validation spec, as
splitting it is more complex and will be done entirely
separately after the basic applicability move is complete.