Start appendix on moved/renamed keywords
Since we're moving (and in a few cases, renaming or splitting)
quite a few keywords, it seems worthwhile to include a brief
accounting of which keywords and why they were moved or
otherwise changed.  This way if someone searches the new text
for the old keywords, at least they'll find something useful.

In the case of "definitions", also advise against redefining
they keyword even though it is no longer technically a part
either spec after being changed to "$defs".