Reorganize assertion/annotation/vocab explanations
This reworks both Overview sections in preparation for
moving the applicators to Core.

The Core overview now covers the generic capabilities
(assertion, annotation, and vocabularies), leaving
the Validation overview to simply reference the concepts.

Rather than noting the standard vocabularies in the
top-level Core overview, they are moved in with the
overview of vocabularies.

Validation and basic meta-data are listed as separate
vocabularies, which is consistent with how they have
been described to date.  Exactly how "format" and
the "content*" keywords are classified into vocabularies
will be finalized in future commits.

The applicators are mentioned as a separate vocabulary in
the core spec, although they are not moved over yet in
this commit.  The descriptive text about applicators which
is removed from the Validation overview in this commit will
be restored to the Core spec when the vocabulary is moved.

There is some duplications as well as the above-mentioned
ommission as a result of this change, which will be addressed
in subsequent commits.