Rework reference sections for new organization
Add an intro paragraph about reference keywords and their
nature as indirect in-place applicators.

"$ref" is no longer really targeted at recursive schemas,
so tweak its wording slightly.

Do a few further improvements to "$recurse" wording, particularly
around root schemas and schema documents.

Correct an apparent typo about a schema "against a schema", this
is almost certainly supposed to be "against an instance", as
the rest of the paragraph talks in terms of instance validation.