Make a Keyword Behavior section
This is purely a reformatting plus the new section tags,
no text has been changed.

Group the Annotation, Assertion, and Application sections.
This new section will get an intro paragraph in the next commit.

Arguably, this has overgrown the "Overview" section, but since
the "Vocabulary" part of the overview is also about to be
expanded, let's wait until then to figure out how to refactor it.