Default behaviors; annotation-dependent results
Describe how default behaviors work in more detail, covering
all three major keyword categories.

Allow keywords to depend on the presence or absence of adjacent
keywords (this is already true with "additionalItems" and "items').

Allow keywords to depend on the annotation results of subschemas
and adjacent keywords.  This is the fundamental mechanism behind
the forthcoming "unevaluatedProperties" and "unevaluatedItems"
keywords.

The behavior of "additionalProperties" and "additionalItems"
can be described with this mechanism (along with the presence
or absence mechanism) as well, which will serve as a proof
of concept of the approach in a subsequent commit.