Define additional* in terms of annotations
This aligns the behavior of "additionalItems" and
"additionalProperties" with the rules given for dependent keywords
in the previous commit.

Specifically, "additionalItems" depends on the presence or
absence of "items", and its newly defined annotation results,
while "additionalProperties" depends on the annotation results
of "properties" and "patternProperties".

The standard explicitly allows implementing these keywords
in other ways (presumably the way implementations currently
handle them) as either an optimization or to support the
keywords in implementations that do not support annotations.

This avoids requiring existing assertion-only validators to
change their behavior.

Note that the rationale for "additionalItems" being ignored when
"items" is absent has been changed to align it with this new
framework.  Doing anything else would have required an absent
keyword to produce annotation results as default behavior,
which leads to very difficult problems as the set of keywords
with such behavior grows.