Remove default value for meta-schema
With default values now having implication for both assertions
and annotation collection, it no longer makes sense to have
a default value for subschemas.

Keywords such as "additionalProperties" now collect annotation
information based on the instance properties to which thier
subschema(s) are applied.  This means that a subschema of true
will cause a property to appear in the annotation, while an
absent subschema will not.  Therefore, a default of true does
not match the annotation collection behavior and is not accurate.

Additionally, as written this has always been problematic, as
it claims a default true value for "not", which would in fact
cause all schemas without "not" to fail.