Rename "dependencies" to "schemaDependencies"
I also noticed that it was in the wrong section.  As an
in-place conditional applicator, it belongs with "if", "then",
and "else".  Not with the object child applicators.

Apparently, I had not made the meta-schema changes at all.
The "$comment" follows the precedent established from leaving
"definitions" in the meta-schema after the rename to "$defs".