Split Keyword Behaviors in to better subsections
This does not change any text at all.  The initial part of
the Keyword Behaviors section is too long.  Put some of the
existing paragraphs into subsections on keyword interactions,
and on default behaviors, which are very distinct topics.

This left the paragraph on extension keywords orphaned, so
I moved it up.  The topic of extension keywords will be
revisited in detail with the vocabulary changes, so it's best
not to worry too much about exactly where this paragraph goes
right now.