Acknowledge that PUT can create
Based on feedback that the current wording, in the context of
all the discussion of POST-to-create-via-collection, can give
the mistaken impression that PUT cannot create.

While we are not responsible for explaining all HTTP semantics,
we can make a bit of an effort here to ensure that we do not
confuse more people.