Note undefined behavior with unknown $ref targets
The most common situation here, which are basically alternate
names for "definitions" or "$defs", will generally work, but
this makes clear that there is no obligation to attempt to
guarantee the proper behavior in such situations.

We might want to narrow this behavior down in the future,
but for now this at least clearly removes any possible burdensome
and totally impractical inferencing requirements.