Make a real, appropriately short, Overview section
The "Keyword Behavior" section is moved, completely unchanged,
to immediately after the "General Considerations", where
it makes quite a bit more sense.  The "General Considerations"
section gives a brief description of the keyword types,
which makes a lot more sense before the detailed description
rather than after it.

The old "Schema Vocabularies" section has mostly been deleted,
as it took up a lot of space without adding much.  Given that
the validation spec now contains four or five vocabularies,
enumerating all of them plus hyper-schema (which is highly
unlikely to reach RFC at the same time as core) seemed
like an overlaod of non-essential information.

The introductory paragraph is still useful, so it has been
spliced into the appropriate location in the
"Meta-Schemas and Vocabularies" section, and the reference
anchor has been given to that section as well.  The contents
of that paragraph are also completely unchanged.