Fix an apparent crediting error
It seems like G. Luff got put here as well as on the hyper-schema
spec when the project restarted, which (based on past credits and
the known authorship of the first post-restart draft) was an
error.

He is still listed in the acknowledgements with other past
major contributors to the overall project.