Make "format" behavior more predictable
By default, a false vocabulary prevents "format" from being validated.
By default, a true vocabulary requires "format" to be validated,
although the degree of validation required remains somewhat vague
at least for this draft.

In both the true and false cases, validation can be toggled on
or off when passing schemas and instances to the implementation
(although in the false case, there is no guarantee at
all that turning on valdiation will produce any validation behavior;
this matches the previous draft's "format" specification).