Clarify annotation collection with applicators
In particular, "contains" collects annotations.  I'm not sure
where the exception came from, as it was not present in the
previous draft (to the extent that that draft mentioned
annotations).

The CREF actually did talk about annotation collection so
apparently this was a bit of a muddled mess anyway.

This makes "contains" behave like all other applicators,
producing both assertion and annotation results.