Consolidate keyword behavior sections
Several sections related to fundamental keyword behavior had
ended up scattered around the document.

The bit about scopes makes more sense near the top of the keyword
behaviors section, as it is really talking about how keywords
interact with the concept of scopes.  It probably does not need
to be in the basic description of schema documents, where it was.

The section on referenced and referencing keywords is closely
associated with applicators, so it becomes a subsection of that.

Finally, the entire section on how to collect annotations gets
moved under the annotations section in keyword behaviors.

This consolidates all generic keyword behavior descriptions
into section 7. Keyword Behaviors.  As you would expect from the name.