Introduce "schema resource" and canonical URIs
This lets us talk more clearly about what an "$id" really is, as
will be seen in the next commit.  It also introduces the idea
of a schema resource having a canonical URI, which is important
for explaining why JSON Pointer fragments relative to parent
base URIs (meaning that they "cross" a subschema "$id") should
have undefined behavior.