Canonical URIs and avoiding shadowed JSON Pointers
This implements the recent decision to strongly discourage using
fragments that cross a base URI change (e.g. an "$id" appearance).

This is done by describing schemas identified by absolute URIs
as resources (per the literal definition of Univeral Resource
Identifier), which was done in the previous commit, and declaring
the URI resulting from the "$id" to be the resource's canonical
URI.

While URIs for subschemas with "$id" and their chidren can be
constructed using the base URI from a parent schema, these
URIs are non-canonical, and their behavior is undefined.  This is
on the grounds that switching between embedding and referencing
schema resources should behave essentially identically.

A difficulty with how annotation collection works in the event
of such as switch is noted in a CREF.