Further clarification of $defs/definitions
Introduce the category of reserved location keywords, which
are basically just $defs and $comment.  This should make it
more clear that you can create keywords of your own like this,
and fits them more nicely into the overall keyword taxonomy.

Part of this change refers to the typical core+validation
meta-schema as the "default meta-schema", which is a change
made in another PR.