More consistent concept of schema resources
This further clarifies the schema resource concept and
how it is used.  It provides an example to simplify
the explanation of how non-canonical URIs can be
problematic.

This also relaxes a restriction on "$schema" that has no functional
impact but avoids requiring implementations to detect and handle
a rather complex case (embedding schema resources with different
"$schema" values).  And also means that embedding can work without
having to change the embedded value by removing the "$schema"
keyword.

Additional best practices for embedding will follow in a separate
commit.