Move core examples/best practices to appendixes
These sections are not normative (RFC 2119 terms will be
removed in a subsequent commit), and moving them to an
appendix reduces the length of the sizable normative
portion of the spec.  And also improves flow, I think.

The examples within core tend to rely on understanding more
of the core keywords than have necessarily been introduced,
so they work better after the entire vocabulary has been
introduced.